# use efficient pattern matcher for wildcard subscriptions