--- a/crossbar/router/observation.py
+++ b/crossbar/router/observation.py
@@ -32,7 +32,7 @@ from __future__ import absolute_import
 import six
 
 from pytrie import StringTrie
-from crossbar.router.pmatcher import PatternMatcher
+from crossbar.router.wildcard import WildcardTrieMatcher
 
 from autobahn import util
 
@@ -153,7 +153,7 @@ class UriObservationMap(object):
         self._observations_prefix = StringTrie()
 
         # map: URI => WildcardUriObservation
-        self._observations_wildcard = PatternMatcher()
+        self._observations_wildcard = WildcardTrieMatcher()
 
         # map: observation ID => UriObservation
         self._observation_id_to_observation = {}