--- a/crossbar/router/test/test_wildcard.py
+++ b/crossbar/router/test/test_wildcard.py
@@ -3,41 +3,59 @@ import unittest
 
 from crossbar.router.wildcard import WildcardMatcher, WildcardTrieMatcher
 
-WILDCARDS = ['a..c', 'a.b.', 'a..']
+WILDCARDS = ['a..c', 'a.b.', 'a..', '.b.']
 MATCHES = {
-    'a': [],
     'a.b': [],
-    'a.b.c': ['a..c', 'a.b.', 'a..'],
+    'a.b.c': ['a..c', 'a.b.', 'a..', '.b.'],
     'a.x.c': ['a..c', 'a..'],
-    'a.b.x': ['a.b.', 'a..'],
+    'a.b.x': ['a.b.', 'a..', '.b.'],
     'a.x.x': ['a..']
 }
 
 
 class TestWildcardMatcher(unittest.TestCase):
-    def test_add(self):
+    def test_setitem(self):
         matcher = WildcardMatcher()
         for w in WILDCARDS:
             matcher[w] = None
-        for w in WILDCARDS:
-            self.assertTrue(w in matcher)
+        self.assertTrue(True)
 
-    def test_del(self):
+    def test_getitem(self):
+        matcher = WildcardMatcher()
+        for i, w in enumerate(WILDCARDS):
+            matcher[w] = i
+        for i, w in enumerate(WILDCARDS):
+            self.assertEqual(matcher[w], i)
+        try:
+            v = matcher['NA']
+        except Exception as e:
+            self.assertTrue(type(e) is KeyError)
+
+    def test_delitem(self):
         matcher = WildcardMatcher()
         for w in WILDCARDS:
             matcher[w] = None
         for w in WILDCARDS:
             del matcher[w]
         for w in WILDCARDS:
-            self.assertTrue(w not in matcher)
+            self.assertFalse(w in matcher)
+
+    def test_contains(self):
+        matcher = WildcardMatcher()
+        for w in WILDCARDS:
+            matcher[w] = None
+        for w in WILDCARDS:
+            self.assertTrue(w in matcher)
+        self.assertFalse('NA' in matcher)
 
     def test_get(self):
         matcher = WildcardMatcher()
         for i, w in enumerate(WILDCARDS):
             matcher[w] = i
         for i, w in enumerate(WILDCARDS):
-            self.assertTrue(matcher.get(w) == i)
+            self.assertEqual(matcher.get(w), i)
         self.assertTrue(matcher.get('NA') is None)
+        self.assertEqual(matcher.get('NA', ''), '')
 
     def test_iter_matches(self):
         matcher = WildcardMatcher()
@@ -45,24 +63,52 @@ class TestWildcardMatcher(unittest.TestCase):
             matcher[w] = w
         for uri, excepted in MATCHES.items():
             s = set(matcher.iter_matches(uri))
-            self.assertTrue(s == set(excepted))
+            self.assertEqual(s, set(excepted))
 
 
 class TestWildcardTrieMatcher(unittest.TestCase):
-    def test_add(self):
+    def test_setitem(self):
         matcher = WildcardTrieMatcher()
         for w in WILDCARDS:
             matcher[w] = None
+        self.assertTrue(True)
+
+    def test_getitem(self):
+        matcher = WildcardTrieMatcher()
         for i, w in enumerate(WILDCARDS):
-            self.assertTrue(matcher.get(w) == i)
+            matcher[w] = i
+        for i, w in enumerate(WILDCARDS):
+            self.assertEqual(matcher[w], i)
+        try:
+            v = matcher['NA']
+        except Exception as e:
+            self.assertTrue(type(e) is KeyError)
+
+    def test_delitem(self):
+        matcher = WildcardTrieMatcher()
+        for w in WILDCARDS:
+            matcher[w] = None
+        for w in WILDCARDS:
+            del matcher[w]
+        for w in WILDCARDS:
+            self.assertFalse(w in matcher)
+
+    def test_contains(self):
+        matcher = WildcardTrieMatcher()
+        for w in WILDCARDS:
+            matcher[w] = None
+        for w in WILDCARDS:
+            self.assertTrue(w in matcher)
+        self.assertFalse('NA' in matcher)
 
     def test_get(self):
         matcher = WildcardTrieMatcher()
         for i, w in enumerate(WILDCARDS):
             matcher[w] = i
         for i, w in enumerate(WILDCARDS):
-            self.assertTrue(matcher.get(w) == i)
+            self.assertEqual(matcher.get(w), i)
         self.assertTrue(matcher.get('NA') is None)
+        self.assertEqual(matcher.get('NA', ''), '')
 
     def test_iter_matches(self):
         matcher = WildcardTrieMatcher()
@@ -70,4 +116,4 @@ class TestWildcardTrieMatcher(unittest.TestCase):
             matcher[w] = w
         for uri, excepted in MATCHES.items():
             s = set(matcher.iter_matches(uri))
-            self.assertTrue(s == set(excepted))
+            self.assertEqual(s, set(excepted))