The current implementation for pattern-based subscriptions is very inefficient.

Here a pull request with a trie-based pattern matcher, tailored for crossbar.io usage. Should be very fast, even with many subscriptions.