@meejah 2015-11-02
Thanks, @yoch!

I won't have time to look closely for a couple days but if we were to merge this, you'd first need to sign a contributer agreement for crossbar: https://github.com/crossbario/crossbar/blob/master/legal/individual_caa.pdf

See here for more: http://crossbar.io/docs/Contributing-to-the-project/