@oberstet 2015-11-02
Thanks, but this already works using a Trie data structure ...