@yoch 2015-11-02
@oberstet : Only prefix subscriptions are managed by StringTrie, but wildcard subscriptions not.