@oberstet 2015-11-02
Correct. And wildcard matching is optimized for the 99% case where you have 1 to a few wildcards only. Do you have performance numbers comparing the existing code with the PR?