@yoch 2015-11-02
I plan to make a benchmark. But trie-based approach performs obviously better with great number of subscriptions.

The real question is how such patterns to consider.