@yoch 2015-11-02
Here benchmark results : 

```text
test on adding 76 wildcards, and searching 100 corresponding uris
	 Old : add	 0.0 s
	 Old : search	 0.015599966049194336 s
	 Old : remove	 0.0 s
	 ----------------------------------------
	 New : add	 0.0 s
	 New : search	 0.0 s
	 New : remove	 0.0 s
	 ----------------------------------------
test on adding 707 wildcards, and searching 1000 corresponding uris
	 Old : add	 0.0 s
	 Old : search	 0.062399864196777344 s
	 Old : remove	 0.0 s
	 ----------------------------------------
	 New : add	 0.015599966049194336 s
	 New : search	 0.015599966049194336 s
	 New : remove	 0.0 s
	 ----------------------------------------
test on adding 3550 wildcards, and searching 5000 corresponding uris
	 Old : add	 0.031200170516967773 s
	 Old : search	 0.2858009338378906 s
	 Old : remove	 0.015599966049194336 s
	 ----------------------------------------
	 New : add	 0.04679989814758301 s
	 New : search	 0.07800102233886719 s
	 New : remove	 0.04679989814758301 s
	 ----------------------------------------
test on adding 7017 wildcards, and searching 10000 corresponding uris
	 Old : add	 0.046801090240478516 s
	 Old : search	 0.6084010601043701 s
	 Old : remove	 0.015599966049194336 s
	 ----------------------------------------
	 New : add	 0.10920000076293945 s
	 New : search	 0.15699982643127441 s
	 New : remove	 0.06240081787109375 s
	 ----------------------------------------
test on adding 14175 wildcards, and searching 20000 corresponding uris
	 Old : add	 0.12479996681213379 s
	 Old : search	 1.2022018432617188 s
	 Old : remove	 0.031199932098388672 s
	 ----------------------------------------
	 New : add	 0.23400115966796875 s
	 New : search	 0.31200098991394043 s
	 New : remove	 0.12479996681213379 s
	 ----------------------------------------
test on adding 35115 wildcards, and searching 50000 corresponding uris
	 Old : add	 0.37540102005004883 s
	 Old : search	 2.9514050483703613 s
	 Old : remove	 0.09360098838806152 s
	 ----------------------------------------
	 New : add	 0.6084010601043701 s
	 New : search	 0.8424010276794434 s
	 New : remove	 0.3432011604309082 s
	 ----------------------------------------
```

http://pastebin.com/dFs5j8Qg