@oberstet 2015-11-02
@yoch alright, Now we are talking;) Thanks backing up with numbers. So there could be something to gain. add/remove gets slower, while search gets faster.

From looking at your test code: this suggests that we should be able to have 2 classes in `pmatcher.py`, one for the old and one for the new matcher. This would then allow to switch between implementations easily and also create unit tests that demonstrate the behavior of the old/new are identical. Maybe we should also rename the file to `wildcard.py` (with 2 exported classes `WildcardTrieMatcher` and `WildcardMatcher` - the native existing one).

If you could come up with a PR that retains both the old/new and let's us switch by changing a single like, I review and merge. Ah, and we need:

* All _current_ Travis tests pass / green. But you already did that. Thanks!
* Please send us a CAA (we only need this once .. pls see @meejah comment)
* Any new files must get the copyright header exactly as in all the other files

Bonus: add new unit tests that demonstrate a) wildcard matching works as expected and/or b) the behavior of the old and new wildcard matcher is identical. That is, unit test that only work on `wildcard.py` (there are a bunch on `observation.py` already).

Thanks for your work!