@yoch 2015-11-03
@oberstet 
Here benchmark modified according to your proposal : http://pastebin.com/1yTq6pDC

It's also possible to speedup `WilcardTrieMatcher.__contains__()` by adding keys into a `set` (at memory cost).

Note also that the longer are the uris, more significant is the difference between the two approaches. In other words, WilcardTrieMatcher can allow more expressive patterns than wamp currently supports.

*I plan to clean up the code and add tests before submitting a PR.*